Convert financial columns from real (float4) to numeric
Money, quantity, and cost-multiplier columns were IEEE-754 floats,
accumulating rounding error in cost rollups, purchasing totals, and
inventory accounting. Migration 0041 converts 74 columns across 27
tables by bucket:

- money    -> numeric(12,2): work order/completion/worklog/part costs,
  labor rates, expenses, cost model rates, cost savings, crew rates,
  equipment values, part/stock unit costs, purchasing amounts
- quantity -> numeric(12,3): stock levels, order quantities (exact
  fractional units)
- cost multipliers -> numeric(6,3) on organizations/equipment/
  cost_model/cost_savings

Deliberately left as real (measurements/statistics where float is the
right type): hours/durations, scores/ratios (roi, quality, confidence,
efficiency), llm token-cost estimates, optimizer planning estimates,
the deprecated inventory_parts table, and all telemetry.

Drizzle columns use numeric({ mode: "number" }) so TS types stay
number end-to-end — tsc passes with zero downstream changes. The
migration is data-driven and idempotent (information_schema guard
converts only columns still real; one rewrite per table per bucket),
so push-bootstrapped databases are untouched and re-runs no-op.

Raw-SQL read paths audited for pg's numeric-as-string behavior:
SUM() aggregates in po-routes/reservations sync job and raw
service-order money SELECTs/RETURNING now cast ::float8 / Number() so
JSON payloads keep returning numbers.

Verified on scratch DB: float values rounded correctly on conversion,
down/up cycle clean, fresh-push and migrate paths converge on
identical column types, dual-schema validator green (numeric and real
normalize to the same bucket on the SQLite side).

+++ server/sync-jobs/reservations.ts
@@ -21,7 +21,7 @@ export async function checkReservationOverflow(orgId: string): Promise<CheckResu
         partId: stock.partId,
         partName: parts.name,
         onHand: stock.quantityOnHand,
-        reserved: sql<number>`COALESCE(SUM(${reservations.quantity}), 0)`,
+        reserved: sql<number>`COALESCE(SUM(${reservations.quantity}), 0)::float8`,
       })
       .from(stock)
       .innerJoin(parts, eq(stock.partId, parts.id))
